+++ src/nucleotides/api/files.sql
@@ -5,22 +5,13 @@ SELECT name FROM file_type
 
 -- name: create-event-file-instance<!
 -- Create a new file_instance and event_file_instance reference
-WITH _existing AS (
-  SELECT id FROM file_instance WHERE sha256 = :sha256
-),
-_created AS (
-  INSERT INTO file_instance (file_type_id, sha256, url)
-  SELECT (SELECT id FROM file_type WHERE name = :type LIMIT 1), :sha256, :url
-  WHERE NOT EXISTS (SELECT 1 FROM _existing)
-  RETURNING id
-),
-_file AS (
-  SELECT * FROM _existing
-  UNION ALL
-  SELECT * FROM _created
+WITH _file AS (
+  SELECT create_file_instance(:sha256, :type, :url) AS id
 )
 INSERT INTO event_file_instance (event_id, file_instance_id)
 VALUES (:event_id, (SELECT id FROM _file))
+
+
 
 
 -- name: get-event-file-instance